+++ src/main.cpp
@@ -27,5 +27,6 @@
 #include "nodes/CCSprite.h"
 #include "nodes/CCSpriteBatchNode.h"
 #include "nodes/CCDirector.h"
-#include "nodes/LevelBrowserLayer.h"
+#include "NodeModding.h"
 #include "BackgroundColors.h"
+#include "texture-loader/PackNode.h"